% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schur_inv.R
\name{schur_inv}
\alias{schur_inv}
\title{The Schur method to get a inverse of a non-symmetric matrix}
\usage{
schur_inv(A, extra_output = FALSE)
}
\arguments{
\item{A}{The matrix}

\item{extra_output}{logical, if you want the factorization of your matrix}
}
\value{

}
\description{
The Schur method to get a inverse of a non-symmetric matrix
}
\examples{
schur_inv(diag(5))
}
