% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theorem_1_9.R
\name{theorem_1_9}
\alias{theorem_1_9}
\title{Using a theorem to get the inverse of a non-symmetric matrix}
\usage{
theorem_1_9(W, extra_output = FALSE)
}
\arguments{
\item{W}{The input matrix}

\item{extra_output}{logical, if you want the factorization of your matrix}
}
\value{

}
\description{
Using a theorem to get the inverse of a non-symmetric matrix
}
